requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
pytest
openai==0.28
python-dotenv
django
flask
ipykernel
pinecone-client==3.0.0 
pinecone-datasets==0.7.0 
langchain-pinecone==0.0.3 
langchain-openai==0.0.7
langchain-community
-e .